% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_students.R
\docType{data}
\name{adm_students}
\alias{adm_students}
\title{Dummy student-level data for Admissions parts B, C, D, F, G, H}
\format{
A data frame with 40 rows (student applicants) and 18 columns
}
\usage{
adm_students
}
\description{
Contains 40 fictional student applicant records with all required data
}
\details{
See complete information by running \code{?create_dummy_data_adm.R}
}
\keyword{datasets}
