% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_upload_format.R
\name{apply_upload_format}
\alias{apply_upload_format}
\title{Shortcut function to turn a dataframe into key-value pairs}
\usage{
apply_upload_format(df)
}
\arguments{
\item{df}{dataframe with upload-compatible column names in upload-compatible order}
}
\value{
a dataframe with one column and upload-compatible rows
}
\description{
Shortcut function to turn a dataframe into key-value pairs
}
