% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/com_cips.R
\docType{data}
\name{com_cips}
\alias{com_cips}
\title{Dummy cip data for Completions functions}
\format{
A data frame with 3 rows and 10 columns
}
\usage{
com_cips
}
\description{
Contains sample values for extra cip codes
}
\details{
See complete information by running \code{?create_dummy_data_com.R}
}
\examples{
\dontrun{produce_com_report(com_student, com_cips)}
}
\keyword{datasets}
