% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/com_students.R
\docType{data}
\name{com_students}
\alias{com_students}
\title{Dummy student data for Completions functions}
\format{
A data frame with 190 rows and 12 columns
}
\usage{
com_students
}
\description{
Contains sample values for students
}
\details{
See complete information by running \code{?create_dummy_data_com.R}
}
\examples{
\dontrun{produce_com_report(com_student, com_cips)}
}
\keyword{datasets}
