% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_autoupload_format_from_aggregate.R
\name{create_autoupload_format_from_aggregate}
\alias{create_autoupload_format_from_aggregate}
\title{Shortcut function to autoformat any file that's already aggregated.
Columns must already be in correct name format.}
\usage{
create_autoupload_format_from_aggregate(df, output_file_name)
}
\arguments{
\item{df}{A dataframe}

\item{output_file_name}{A string containing a file name for your results}
}
\value{
A txt file at the path of your choice
}
\description{
Shortcut function to autoformat any file that's already aggregated.
Columns must already be in correct name format.
}
