% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompletionsStartingDf_DummyData.R
\name{create_dummy_data_comp}
\alias{create_dummy_data_comp}
\title{Create dummy data for testing the completions functions}
\usage{
create_dummy_data_comp(df_type = "student")
}
\arguments{
\item{df_type}{a string 'student' to get the main df needed, 'cip' to get extracips}
}
\value{
a dataframe ready for the rest of the comp scripts
}
\description{
to do: save this out into the package and make it accessible as package data
}
