% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dummy_data_ef1.R
\name{create_dummy_data_ef1}
\alias{create_dummy_data_ef1}
\title{Make Fall Enrollment Dummy Data}
\usage{
create_dummy_data_ef1(n = 100, seed = 1234)
}
\arguments{
\item{n}{A number}

\item{seed}{A number}
}
\value{
A text file
}
\description{
Make Fall Enrollment Dummy Data
}
