% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dummy_data_gr.R
\name{create_dummy_data_gr}
\alias{create_dummy_data_gr}
\title{Create dummy data for testing the Grad Rates functions}
\usage{
create_dummy_data_gr(n = 1000, seed = 4567)
}
\arguments{
\item{n}{Number of rows of data to synthesize}

\item{seed}{Seed number to supply to \code{"set.seed()"} to replicate of results}
}
\value{
a dataframe ready for the rest of the Grad Rates functions
}
\description{
Creates dummy data for testing the Grad Rates functions
}
