% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dummy_data_gr200.R
\name{create_dummy_data_gr200}
\alias{create_dummy_data_gr200}
\title{Create dummy data for testing the Grad Rates 200 function}
\usage{
create_dummy_data_gr200(n = 1000)
}
\arguments{
\item{n}{A number that will be used as the length of the data frame}
}
\value{
a dataframe ready for the rest of the Grad Rates 200 functions
}
\description{
Dummy data for Grad Rates 200 testing
}
\examples{
set.seed(4567)

#default creates 1000 students
graduates <- create_dummy_data_gr200()
more_graduates <- create_dummy_data_gr200(n = 100)

}
