% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HumanResources_DummyData.R
\name{create_dummy_data_hr}
\alias{create_dummy_data_hr}
\title{Create dummy data for testing the hr functions}
\usage{
create_dummy_data_hr()
}
\value{
a dataframe ready for the rest of the hr scripts
}
\description{
to do: save this out into the package and make it accessible as package data
}
