% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dummy_data_om.R
\name{create_dummy_data_om}
\alias{create_dummy_data_om}
\title{Create dummy data for testing the outcome measures functions}
\usage{
create_dummy_data_om()
}
\value{
a dataframe ready for the rest of the om scripts
}
\description{
Creates a prepared dataframe to test scripts related to IPEDS Outcome Measures reporting.
Produces either a student/status dataframe
}
\details{
remember: want to save this data out into the package so it's available
}
\note{
The final dataset has 20 students covering most statuses
}
