% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_OM.R
\docType{data}
\name{dummy_OM}
\alias{dummy_OM}
\title{Dummy data for test OM functions}
\format{
A data frame with 20 rows and 9 columns
}
\usage{
dummy_OM
}
\description{
Dummy data for test OM functions
}
\keyword{datasets}
