% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e1d_instr.R
\docType{data}
\name{e1d_instr}
\alias{e1d_instr}
\title{Dummy aggregated data for 12 Month Enrollment part B}
\format{
A data frame with 1 row and 5 columns
}
\usage{
e1d_instr
}
\description{
Contains sample values for credit hours generated and doctors-professional FTE
}
\details{
See complete information by running \code{?create_dummy_data_e1d.R}
}
\examples{
\dontrun{produce_e1d_report(e1d_student, e1d_instr)}
}
\keyword{datasets}
