% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e1d_student.R
\docType{data}
\name{e1d_student}
\alias{e1d_student}
\title{Dummy student-level data for 12 Month Enrollment parts A and C}
\format{
A data frame with 100 rows (students) and 11 columns
}
\usage{
e1d_student
}
\description{
Contains 100 fictional student records with all required data
}
\details{
See complete information by running \code{?create_dummy_data_e1d.R}
}
\keyword{datasets}
