% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef1_retention.R
\docType{data}
\name{ef1_retention}
\alias{ef1_retention}
\title{Dataframe of dummy retention data for Fall Enrollment scripts}
\format{
A data frame with 2 rows and 6 columns
}
\usage{
ef1_retention
}
\description{
This data provides aggregated counts in a dataframe suitable for use in the retention component of the Fall Enrollment survey.
}
\keyword{datasets}
