% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef1_students.R
\docType{data}
\name{ef1_students}
\alias{ef1_students}
\title{Dummy student data for Fall Enrollment scripts}
\format{
A data frame with 100 rows and 25 columns
}
\usage{
ef1_students
}
\description{
Using the default number of students, this data
  provides a population that touches most available categories of student
  reporting. Some columns use only a selection of possible values to reduce
  complexity.
}
\note{
To recreate the saved dataframe exactly, use seed 1234 with 100 students.
}
\keyword{datasets}
