% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr200_students.R
\docType{data}
\name{gr200_students}
\alias{gr200_students}
\title{Dummy student data for Graduation Rates 200 functions}
\format{
A data frame with 1000 rows and 5 columns
}
\usage{
gr200_students
}
\description{
Contains sample values for students
}
\details{
See complete information by running \code{?create_dummy_data_gr200.R}
}
\keyword{datasets}
