% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_staff.R
\docType{data}
\name{hr_staff}
\alias{hr_staff}
\title{Dummy staff data for Human Resources functions}
\format{
A data frame with 3600 rows and 13 columns
}
\usage{
hr_staff
}
\description{
Contains sample values for staff
}
\details{
See complete information by running \code{?create_dummy_data_hr.R}
}
\examples{
\dontrun{produce_hr_report(hr_staff)}
}
\keyword{datasets}
