% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPEDSuploadables-package.R
\name{IPEDSuploadables}
\alias{IPEDSuploadables}
\title{\code{IPEDSuploadables} package}
\description{
Tools to assist data formatting for upload to IPEDS surveys
}
\details{
See the README on
\href{https://github.com/AlisonLanski/IPEDSuploadables#readme}{GitHub}
or view documentation at \href{https://alisonlanski.github.io/IPEDSuploadables/}{the pkgdown site}
}
