% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_adm_part_B.R
\name{make_adm_part_B}
\alias{make_adm_part_B}
\title{Make Admissions Part B (First-time student demographics)}
\usage{
make_adm_part_B(df)
}
\arguments{
\item{df}{A dataframe of applicant information}
}
\value{
Admissions Part B data with the required IPEDS structure
}
\description{
Aggregates First-Time student applicants by sex, race/ethnicity.
}
