% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_adm_part_G.R
\name{make_adm_part_G}
\alias{make_adm_part_G}
\title{Make Admissions Part G (Transfer sex unknown)}
\usage{
make_adm_part_G(df)
}
\arguments{
\item{df}{A dataframe of applicant information}
}
\value{
Admissions Part G data with the required IPEDS structure
}
\description{
Counts the number of unknown sex for Transfer students
}
