% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_com_part_B.R
\name{make_com_part_B}
\alias{make_com_part_B}
\title{Make Completions Part B}
\usage{
make_com_part_B(df, extracips = NULL)
}
\arguments{
\item{df}{A dataframe}

\item{extracips}{A dataframe}
}
\value{
A text file
}
\description{
Make Completions Part B
}
