% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_com_part_D.R
\name{make_com_part_D}
\alias{make_com_part_D}
\title{Make Completions Part D}
\usage{
make_com_part_D(df, extracips = NULL)
}
\arguments{
\item{df}{A dataframe of student/degree information}

\item{extracips}{A dataframe of cips offered by the institution but not in \code{'df'}}
}
\value{
A dataframe with the required IPEDS structure for this survey part
}
\description{
Make Completions Part D
}
