% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_com_part_E.R
\name{make_com_part_E}
\alias{make_com_part_E}
\title{Make Completions Part E (gender details)}
\usage{
make_com_part_E(
  df,
  ugender = lifecycle::deprecated(),
  ggender = lifecycle::deprecated()
)
}
\arguments{
\item{df}{A dataframe of student/degree information}

\item{ugender}{`r lifecycle::badge("deprecated")` A boolean: TRUE means you are collecting and able to report
"another gender" for undergraduate completers, even if you have no (or few)
such students. Set as FALSE if necessary. **Starting in 2025-2026, this argument will be ignored by later
code.**}

\item{ggender}{`r lifecycle::badge("deprecated")` A boolean: TRUE means you are collecting and able to report
"another gender" for graduate completers, even if you have no (or few) such
students. Set as FALSE if necessary. **Starting in 2025-2026, this argument will be ignored by later
code.**}
}
\value{
A dataframe with the required IPEDS structure for this survey part
}
\description{
Make Completions Part E (gender details)
}
