% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_completions.R
\name{make_completions}
\alias{make_completions}
\title{Make all of Completions}
\usage{
make_completions(df, extracips = NULL)
}
\arguments{
\item{df}{A dataframe}

\item{extracips}{A dataframe}
}
\value{
A text file
}
\description{
Make all of Completions
}
