% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_e12_part_A.R, R/make_e1d_part_A.R
\name{make_e1d_part_A}
\alias{make_e1d_part_A}
\title{Make 12 Month Enrollment Part A}
\usage{
make_e1d_part_A(df)

make_e1d_part_A(df)
}
\arguments{
\item{df}{A dataframe of student/degree information}
}
\value{
A dataframe with the required IPEDS structure for this survey part

A dataframe with the required IPEDS structure for this survey part
}
\description{
Make 12 Month Enrollment Part A

Make 12 Month Enrollment Part A
}
