% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_e1d_part_B.R
\name{make_e1d_part_B}
\alias{make_e1d_part_B}
\title{Make 12 Month Enrollment Part B}
\usage{
make_e1d_part_B(df)
}
\arguments{
\item{df}{A dataframe with summarized credit hours and student information}
}
\value{
A dataframe with the required IPEDS structure for this survey part
}
\description{
Make 12 Month Enrollment Part B
}
