% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_e1d_part_C.R
\name{make_e1d_part_C}
\alias{make_e1d_part_C}
\title{Make 12 Month Enrollment Part C}
\usage{
make_e1d_part_C(df, output = "part", format = "both")
}
\arguments{
\item{df}{A dataframe of student/degree information}

\item{output}{A string (\code{"part"}, \code{"full"}, or \code{"both"})}

\item{format}{A string (\code{"uploadable"}, \code{"readable"}, or \code{"both"})}
}
\value{
A text file
}
\description{
Make 12 Month Enrollment Part C
}
