% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_e1d_part_D.R
\name{make_e1d_part_D}
\alias{make_e1d_part_D}
\title{Make 12 Month Enrollment Part D (gender details)}
\usage{
make_e1d_part_D(
  df,
  ugender = lifecycle::deprecated(),
  ggender = lifecycle::deprecated()
)
}
\arguments{
\item{df}{A dataframe of student/degree information}

\item{ugender}{`r lifecycle::badge("deprecated")` A boolean: TRUE means you are collecting and able to report
"another gender" for undergraduate students, even if you have no (or few)
such students. Set as FALSE if necessary. **Starting in 2025-2026, this argument will be ignored by later
code.**}

\item{ggender}{`r lifecycle::badge("deprecated")` A boolean: TRUE means you are collecting and able to report
"another gender" for graduate students, even if you have no (or few) such
students. Set as FALSE if necessary. **Starting in 2025-2026, this argument will be ignored by later
code.**}
}
\value{
A dataframe with the required IPEDS structure for this survey part
}
\description{
Make 12 Month Enrollment Part D (gender details)
}
