% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_e1d_part_E.R
\name{make_e1d_part_E}
\alias{make_e1d_part_E}
\title{Make 12 Month Enrollment Part E}
\usage{
make_e1d_part_E(df)
}
\arguments{
\item{df}{A dataframe of student/degree information}
}
\value{
A dataframe with the required IPEDS structure for this survey part
}
\description{
R/E and Gender counts for dual enrollment (high school students)
}
