% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_e12_part_F.R, R/make_e1d_part_F.R
\name{make_e1d_part_F}
\alias{make_e1d_part_F}
\title{Make 12 Month Enrollment Part F}
\usage{
make_e1d_part_F(df)

make_e1d_part_F(df)
}
\arguments{
\item{df}{A dataframe of student/degree information}
}
\value{
A dataframe with the required IPEDS structure for this survey part

A dataframe with the required IPEDS structure for this survey part
}
\description{
Flag questions about high school students enrolled for credit

Flag questions about high school students enrolled for credit
}
