% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ef1_part_B.R
\name{make_ef1_part_B}
\alias{make_ef1_part_B}
\title{Make Fall Enrollment Part B}
\usage{
make_ef1_part_B(df, output = "part", format = "both")
}
\arguments{
\item{df}{A dataframe of student information}

\item{output}{A string (\code{"part"}, \code{"full"}, or \code{"both"})}

\item{format}{A string (\code{"uploadable"}, \code{"readable"}, or \code{"both"})}
}
\value{
A text file
}
\description{
Student Counts by Age/gender
}
