% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ef1_part_C.R
\name{make_ef1_part_C}
\alias{make_ef1_part_C}
\title{Make Fall Enrollment Part C}
\usage{
make_ef1_part_C(df)
}
\arguments{
\item{df}{A dataframe of student/degree information}
}
\value{
A text file
}
\description{
State of origin for first time students
}
