% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ef1_part_D.R
\name{make_ef1_part_D}
\alias{make_ef1_part_D}
\title{Make Fall Enrollment Part D}
\usage{
make_ef1_part_D(df)
}
\arguments{
\item{df}{A dataframe of student/degree information}
}
\value{
A text file
}
\description{
Count of new non-degree students
}
