% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ef1_part_E.R
\name{make_ef1_part_E}
\alias{make_ef1_part_E}
\title{Make Fall Enrollment Part E}
\usage{
make_ef1_part_E(df)
}
\arguments{
\item{df}{A dataframe of student/degree information}
}
\value{
A text file
}
\description{
Retention counts
}
