% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ef1_part_F.R
\name{make_ef1_part_F}
\alias{make_ef1_part_F}
\title{Make Fall Enrollment Part F}
\usage{
make_ef1_part_F(df)
}
\arguments{
\item{df}{A dataframe (either "students" or "retention") as a unitid source}
}
\value{
A text file
}
\description{
Student Faculty Ratio
}
