% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ef1_part_G.R
\name{make_ef1_part_G}
\alias{make_ef1_part_G}
\title{Make Fall Enrollment Part G}
\usage{
make_ef1_part_G(df)
}
\arguments{
\item{df}{A dataframe of student/degree information}
}
\value{
A dataframe with the required IPEDS structure for this survey part
}
\description{
Distance Ed counts
}
