% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ef1_part_H.R
\name{make_ef1_part_H}
\alias{make_ef1_part_H}
\title{Make Fall Enrollment Part H (gender details)}
\usage{
make_ef1_part_H(df, ugender, ggender)
}
\arguments{
\item{df}{A dataframe of student enrollment information}

\item{ugender}{A boolean: TRUE means you are collecting and able to report
"another gender" for undergraduate completers, even if you have no (or few)
such students. Set as FALSE if necessary}

\item{ggender}{A boolean: TRUE means you are collecting and able to report
"another gender" for graduate completers, even if you have no (or few) such
students. Set as FALSE if necessary}
}
\value{
A text file
}
\description{
Make Fall Enrollment Part H (gender details)
}
