% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gr200.R
\name{make_gr200}
\alias{make_gr200}
\title{Make Graduation Rates 200}
\usage{
make_gr200(df, output = "part", format = "both")
}
\arguments{
\item{df}{A dataframe of student/degree information}

\item{output}{A string (\code{"part"}, \code{"full"}, or \code{"both"})}

\item{format}{A string (\code{"uploadable"}, \code{"readable"}, or \code{"both"})}
}
\value{
A text file
}
\description{
Make Graduation Rates 200
}
