% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gr2002.R
\name{make_gr2002}
\alias{make_gr2002}
\title{Make Graduation Rates 200}
\usage{
make_gr2002(df)
}
\arguments{
\item{df}{A dataframe of student/degree information}
}
\value{
A dataframe with the required IPEDS structure for this survey part
}
\description{
Make Graduation Rates 200
}
