% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gr_part_B.R
\name{make_gr_part_B}
\alias{make_gr_part_B}
\title{Make Graduation Rates Part B}
\usage{
make_gr_part_B(df, output = "part", format = "both")
}
\arguments{
\item{df}{A dataframe of student/degree information}

\item{output}{A string (\code{"part"}, \code{"full"}, or \code{"both"})}

\item{format}{A string (\code{"uploadable"}, \code{"readable"}, or \code{"both"})}
}
\value{
A text file
}
\description{
Make Graduation Rates Part B
}
