% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gr_part_C.R
\name{make_gr_part_C}
\alias{make_gr_part_C}
\title{Make Graduation Rates Part C}
\usage{
make_gr_part_C(df, output = "part", format = "both")
}
\arguments{
\item{df}{A dataframe of student/degree information}

\item{output}{A string (\code{"part"}, \code{"full"}, or \code{"both"})}

\item{format}{A string (\code{"uploadable"}, \code{"readable"}, or \code{"both"})}

\item{cohort_year}{A numerical value for the fall cohort}
}
\value{
A text file
}
\description{
Make Graduation Rates Part C
}
