% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gr_part_E.R
\name{make_gr_part_E}
\alias{make_gr_part_E}
\title{Make Graduation Rates Part E (gender details)}
\usage{
make_gr_part_E(df, ugender)
}
\arguments{
\item{df}{A dataframe of student/degree information for unduplicated undergraduates}

\item{ugender}{A boolean: TRUE means you are collecting and able to report
"another gender" for undergraduate students. Set as FALSE if necessary}
}
\value{
A df aggregated for the survey part
}
\description{
Make Graduation Rates Part E (gender details)
}
