% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hr_part_A1.R
\name{make_hr_part_A1}
\alias{make_hr_part_A1}
\title{Produce IPEDS HR Part A1}
\usage{
make_hr_part_A1(df, output = "part", format = "both")
}
\arguments{
\item{df}{a dataframe}

\item{output}{A string (\code{"part"}, \code{"full"}, or \code{"both"})}

\item{format}{A string (\code{"uploadable"}, \code{"readable"}, or \code{"both"})}
}
\value{
a txt file
}
\description{
Part A1 --- COUNT of FT INSTRUCTIONAL staff by tenure status, academic rank, and race/ethnicity/gender
}
