% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hr_part_A2.R
\name{make_hr_part_A2}
\alias{make_hr_part_A2}
\title{Make Human Resources Part A2}
\usage{
make_hr_part_A2(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
A dataframe with the required IPEDS structure for this survey part
}
\description{
Part A2 --- COUNT of FT instructional staff by tenure status, medical school, and function
}
