% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hr_part_B2.R
\name{make_hr_part_B2}
\alias{make_hr_part_B2}
\title{Make Human Resources Part B2}
\usage{
make_hr_part_B2(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
A dataframe with the required IPEDS structure for this survey part
}
\description{
Part B2 --- Full-time non-instructional staff by tenure, medical school, and occupational category
}
