% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hr_part_B3.R
\name{make_hr_part_B3}
\alias{make_hr_part_B3}
\title{Produce IPEDS HR Part B3}
\usage{
make_hr_part_B3(df, output = "part")
}
\arguments{
\item{df}{a dataframe}

\item{output}{a string (\code{"part"}, \code{"full"}, or \code{"both"})}
}
\value{
a txt file
}
\description{
Part B3 --- Full-time non-instructional staff by medical school, and occupational category
}
