% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hr_part_D1.R
\name{make_hr_part_D1}
\alias{make_hr_part_D1}
\title{Produce IPEDS HR Part D1}
\usage{
make_hr_part_D1(df, output = "part")
}
\arguments{
\item{df}{a dataframe}

\item{output}{a string (\code{"part"}, \code{"full"}, or \code{"both"})}
}
\value{
a txt file
}
\description{
Part D1 --- Part-time staff by occupational category
}
