% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hr_part_D2.R
\name{make_hr_part_D2}
\alias{make_hr_part_D2}
\title{Make Human Resources Part D2}
\usage{
make_hr_part_D2(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
A dataframe with the required IPEDS structure for this survey part
}
\description{
Part D2 --- Graduate assistants by occupational category and race/ethnicity/gender
}
