% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hr_part_D3.R
\name{make_hr_part_D3}
\alias{make_hr_part_D3}
\title{Produce IPEDS HR Part D3}
\usage{
make_hr_part_D3(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
a txt file
}
\description{
Part D3 --- Part-time staff by tenure, medical school, and occupational category
}
