% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hr_part_D4.R
\name{make_hr_part_D4}
\alias{make_hr_part_D4}
\title{Make Human Resources Part D4}
\usage{
make_hr_part_D4(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
A dataframe with the required IPEDS structure for this survey part
}
\description{
Part D4 --- Part-time Non-instructional staff by medical school, and occupational category
}
