% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hr_part_G1.R
\name{make_hr_part_G1}
\alias{make_hr_part_G1}
\title{Produce IPEDS HR Part G1}
\usage{
make_hr_part_G1(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
a txt file
}
\description{
Part G1 --- Salaries of INSTRUCTIONAL staff
}
