% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hr_part_G2.R
\name{make_hr_part_G2}
\alias{make_hr_part_G2}
\title{Make Human Resources Part G2}
\usage{
make_hr_part_G2(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
A dataframe with the required IPEDS structure for this survey part
}
\description{
Part G2 --- Salaries of non-instructional staff
}
