% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hr_part_H1.R
\name{make_hr_part_H1}
\alias{make_hr_part_H1}
\title{Produce IPEDS HR Part H1}
\usage{
make_hr_part_H1(df, output = "part")
}
\arguments{
\item{df}{a dataframe}

\item{output}{a string (\code{"part"}, \code{"full"}, or \code{"both"})}
}
\value{
a txt file
}
\description{
Part H1 --- Full-time new hire instructional staff by tenure status and race/ethnicity/gender
}
