% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hr_part_H2.R
\name{make_hr_part_H2}
\alias{make_hr_part_H2}
\title{Produce IPEDS HR Part H2}
\usage{
make_hr_part_H2(df, output = "part")
}
\arguments{
\item{df}{a dataframe}

\item{output}{a string (\code{"part"}, \code{"full"}, or \code{"both"})}
}
\value{
a txt file
}
\description{
Part H2 --- New hires by occupational category, Race/Ethnicity/Gender
}
