% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_make_pretty_files.R
\name{make_pretty_files}
\alias{make_pretty_files}
\title{Make Pretty Files}
\usage{
make_pretty_files(df)
}
\arguments{
\item{df}{a data frame returned from IPEDS}
}
\description{
Creates an easy to read output file.
}
\examples{
\dontrun{df <- make_pretty_files(df = partA)}

}
