% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_students.R
\docType{data}
\name{om_students}
\alias{om_students}
\title{Dummy data for Outcome Measures functions}
\format{
A data frame with 20 rows and 9 columns
}
\usage{
om_students
}
\description{
Contains sample values for students
}
\details{
See complete information by running \code{?create_dummy_data_om.R}
}
\keyword{datasets}
