% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part_outputs.R
\docType{data}
\name{part_outputs}
\alias{part_outputs}
\title{Output files from the make functions using the package dummy data (for testing)}
\format{
A list of dataframes: one dataframe per make script
}
\usage{
part_outputs
}
\description{
Contains a list of output files from each "make" script which can be called by testthat tests to ensure refactoring doesn't change output inappropriately
}
\details{
General list-creation code is commented out in the test-part-outputs.R script; uncomment and run/add as needed if a df needs a legitimate update (e.g. IPEDS rules change, dummy data includes new use case)
}
\examples{
\dontrun{part_outputs$com_partD}
}
\keyword{internal}
