% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_com_data_files.R
\name{prep_com_data_files}
\alias{prep_com_data_files}
\title{Some initial recodes for Completions}
\usage{
prep_com_data_files(df)
}
\arguments{
\item{df}{a dataframe of student level data}
}
\value{
df
}
\description{
Some initial recodes for Completions
}
