% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_ef1_data_frame.R
\name{prep_ef1_data_frame}
\alias{prep_ef1_data_frame}
\title{Some initial recodes for Fall Enrollment}
\usage{
prep_ef1_data_frame(df)
}
\arguments{
\item{df}{a dataframe of student level data}
}
\value{
df
}
\description{
Some initial recodes for Fall Enrollment
}
