% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_hr_data_frame.R
\name{prep_hr_data_frame}
\alias{prep_hr_data_frame}
\title{Prepare dataframe for IPEDS HR script}
\usage{
prep_hr_data_frame(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
a dataframe
}
\description{
Prepare dataframe for IPEDS HR script
}
