% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_om_data_frame.R
\name{prep_om_data_frame}
\alias{prep_om_data_frame}
\title{Some initial recoding for OutcomeMeasures}
\usage{
prep_om_data_frame(df)
}
\arguments{
\item{df}{a dataframe of student level data}
}
\value{
A dataframe ready for the make_om scripts
}
\description{
Some initial recoding for OutcomeMeasures
}
