% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_adm_report.R
\name{produce_adm_report}
\alias{produce_adm_report}
\title{Shortcut function to do all steps to produce a report}
\usage{
produce_adm_report(df, ptype = 7, part = "ALL", format = "uploadable")
}
\arguments{
\item{df}{A dataframe set up according to the readme}

\item{ptype}{(Optional) An integer [1-9] indicating which calculation method to use for test score percentiles. The default value within R and here is 7. To see details, run \code{?quantile} and scroll down to "Type".}

\item{part}{A string with what part of the report you want to produce: 'all', 'A', etc.}

\item{format}{A string (\code{"uploadable"} will produce a properly formatted
upload file. \code{"readable"} will produce a csv of the upload file (only
works for one part at a time). \code{"both"} will provide both options, but
only works with one part at a time.}
}
\value{
A txt or csv file at the path of your choice
}
\description{
Shortcut function to do all steps to produce a report
}
