% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_e1d_report.R
\name{produce_e1d_report}
\alias{produce_e1d_report}
\title{Shortcut function to do all steps to produce a 12 month Enrollment report}
\usage{
produce_e1d_report(df, hrs, part = "ALL")
}
\arguments{
\item{df}{A dataframe set up according to the readme for students}

\item{hrs}{A dataframe set up according to the readme for instructional activity}

\item{part}{A string with what part of the reprt you want to produce: 'all', 'A', etc.}
}
\value{
A txt file at the path of your choice
}
\description{
Shortcut function to do all steps to produce a 12 month Enrollment report
}
