% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_hr_report.R
\name{produce_hr_report}
\alias{produce_hr_report}
\title{Shortcut function to do all steps to produce a report}
\usage{
produce_hr_report(df, part = "all")
}
\arguments{
\item{df}{a dataframe set up according to the readme}

\item{part}{a string with what part of the report you want to produce 'all', 'A1', etc.}
}
\value{
a txt file at the path of your choice
}
\description{
Shortcut function to do all steps to produce a report
}
