% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_reg.R
\name{recode_reg}
\alias{recode_reg}
\title{Recode race/ethnicity and gender}
\usage{
recode_reg(df)
}
\arguments{
\item{df}{A dataframe of student information}
}
\value{
A text file
}
\description{
Recode race/ethnicity and gender
}
