% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_ipeds_unitid.R
\name{set_ipeds_unitid}
\alias{set_ipeds_unitid}
\title{Establish your insititution's Unitid for the IPEDS script}
\usage{
set_ipeds_unitid()
}
\value{
a numeric unitid
}
\description{
Establish your insititution's Unitid for the IPEDS script
}
