% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_report_path.R
\name{set_report_path}
\alias{set_report_path}
\title{Set the path for where the reports will be saved to.}
\usage{
set_report_path()
}
\value{
path
}
\description{
Set the path for where the reports will be saved to.
}
