% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs_ADM.R
\docType{data}
\name{specs_ADM}
\alias{specs_ADM}
\title{Table of data requirements for the Admissions starting dataframe}
\format{
A data frame with one row per required fact and 3 columns
}
\usage{
specs_ADM
}
\description{
Table of data requirements for the Admissions starting dataframe
}
\keyword{datasets}
