% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs_COM.R
\docType{data}
\name{specs_COM}
\alias{specs_COM}
\title{Table of data requirements for Completions starting dataframe}
\format{
A data frame with 19 rows and 4 columns
}
\usage{
specs_COM
}
\description{
Table of data requirements for Completions starting dataframe
}
\keyword{datasets}
