% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs_E1D.R
\docType{data}
\name{specs_E1D}
\alias{specs_E1D}
\title{Table of data requirements for 12 Month Enrollment starting dataframes}
\format{
A data frame with 19 rows and 4 columns
}
\usage{
specs_E1D
}
\description{
Table of data requirements for 12 Month Enrollment starting dataframes
}
\keyword{datasets}
