% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs_EF1.R
\docType{data}
\name{specs_EF1}
\alias{specs_EF1}
\title{Table of data requirements for Fall Enrollment starting dataframes}
\format{
A data frame with 23 rows and 4 columns
}
\usage{
specs_EF1
}
\description{
Table of data requirements for Fall Enrollment starting dataframes
}
\keyword{datasets}
