% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs_GR.R
\docType{data}
\name{specs_GR}
\alias{specs_GR}
\title{Table of data requirements for Grad Rates starting dataframe}
\format{
A data frame with 8 rows and 3 columns
}
\usage{
specs_GR
}
\description{
Table of data requirements for Grad Rates starting dataframe
}
\keyword{datasets}
