% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs_GR200.R
\docType{data}
\name{specs_GR200}
\alias{specs_GR200}
\title{Table of data requirements for Grad Rates 200 starting dataframe}
\format{
A data frame with 5 rows and 3 columns
}
\usage{
specs_GR200
}
\description{
Table of data requirements for Grad Rates 200 starting dataframe
}
\keyword{datasets}
