% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs_HR.R
\docType{data}
\name{specs_HR}
\alias{specs_HR}
\title{Table of data requirements for HR starting datagfame}
\format{
A data frame with 13 rows and 3 columns
}
\usage{
specs_HR
}
\description{
Table of data requirements for HR starting datagfame
}
\keyword{datasets}
