% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs_OM.R
\docType{data}
\name{specs_OM}
\alias{specs_OM}
\title{Table of data requirements for OM starting dataframe}
\format{
A data frame with 9 rows and 3 columns
}
\usage{
specs_OM
}
\description{
Table of data requirements for OM starting dataframe
}
\keyword{datasets}
