% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_report_csv.R
\name{write_report_csv}
\alias{write_report_csv}
\title{Write the prepared data to a csv file}
\usage{
write_report_csv(df, survey, part, output_path)
}
\arguments{
\item{df}{a dataframe (prepared via the 'make' scripts)}

\item{survey}{a string (which IPEDS survey)}

\item{part}{a string (which upload part of the survey)}

\item{output_path}{a path (which folder the report should go in)}
}
\value{
a csv file (at the path location)
}
\description{
Write the prepared data to a csv file
}
\note{
All arguments for this function are required.
  The dataframe must have the key as the column name (with appropriate
  capitalization) and the value in the cells
}
