% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dummy_data_adm.R
\name{create_dummy_data_adm}
\alias{create_dummy_data_adm}
\title{Create dummy data for testing the Admissions functions}
\usage{
create_dummy_data_adm(seed = 4567)
}
\arguments{
\item{seed}{Number to set a seed to randomize exam scores}
}
\value{
a dataframe of 40 applicants ready for the rest of the Admissions functions
}
\description{
Creates dummy data for testing the Admissions functions
}
\examples{

#use default seed
admissions <- create_dummy_data_adm()

#use custom seed
admissions_scores <- create_dummy_data_adm(seed = 123456)
}
