% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dummy_data_comp.R
\name{create_dummy_data_comp}
\alias{create_dummy_data_comp}
\title{Create dummy data for testing the completions functions}
\usage{
create_dummy_data_comp(df_type = "student")
}
\arguments{
\item{df_type}{a string: 'student' to get the main df needed, 'cip' to get extracips}
}
\value{
a dataframe ready for the rest of the comp scripts
}
\description{
Creates a prepared dataframe to test scripts related to IPEDS Completions reporting.
Produces either a student/degree dataframe or a dataframe of cips previously reported but not in the current student data,
depending on the argument you select
}
\details{
remember: want to save this data out into the package so it's available
}
\note{
The final dataset has 100 students with 190 majors.
60 students have 1 major for 1 degree
8 students have 2 majors for 1 degree
17 students have 3 majors (on one degree or as dual-degree)
12 students have 4 majors (as dual-degrees; some have one as an MBA)
3 students have 5 majors (dual-degree and 4 majors and an MBA)
Note: 1 student has a faulty birthdate; this will show the warning "1 failed to parse"
I use the same seed for each grouping,
to ensure that the same students keep getting more things added
One program/level combination is flagged as distance education
With the seed and probs, no associates degrees end up in the data (don't worry!)
To fully process completions, we will need to include an example
of a CIP code that is a possible major but has no completers
and a CIP code in an award level that is possible but has no completers
This is the second piece of dummy df produced
}
