% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ipeds_unitid.R
\name{get_ipeds_unitid}
\alias{get_ipeds_unitid}
\title{Grab institution's UNITID from supplied data to populate missing-data rows}
\usage{
get_ipeds_unitid(df)
}
\arguments{
\item{df}{a dataframe with ipeds data and one unitid}
}
\value{
a character unitid
}
\description{
Grab institution's UNITID from supplied data to populate missing-data rows
}
