% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_students.R
\docType{data}
\name{gr_students}
\alias{gr_students}
\title{Dummy student data for the Graduation Rates scripts}
\format{
A data frame with 101 rows and 13 columns
}
\usage{
gr_students
}
\description{
Dummy student data for the Graduation Rates scripts
}
\details{
Includes only 3 Race/Ethnicity categories [6, 7, 8] for simpler
  code; one student (a program-switcher) has a 4th category [1] for easy
  tracking
}
\keyword{datasets}
