% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPEDSuploadables-package.R
\docType{package}
\name{IPEDSuploadables-package}
\alias{IPEDSuploadables-package}
\title{IPEDSuploadables: Transforms Institutional Data into Text Files for IPEDS Automated Import/Upload}
\description{
Starting from user-supplied institutional data, these scripts transform, aggregate, and reshape the information to produce key-value pair data files that are able to be uploaded to IPEDS (Integrated Postsecondary Education Data System) through their submission portal \url{https://surveys.nces.ed.gov/ipeds/}. Starting data specifications can be found in the vignettes. Final files are saved locally to a location of the user's choice. User-friendly readable files can also be produced for purposes of data review and validation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AlisonLanski/IPEDSuploadables}
  \item \url{https://alisonlanski.github.io/IPEDSuploadables/}
  \item Report bugs at \url{https://github.com/AlisonLanski/IPEDSuploadables/issues}
}

}
\author{
\strong{Maintainer}: Alison Lanski \email{alanski@nd.edu}

Authors:
\itemize{
  \item Shiloh Fling \email{shilohfling@gmail.com}
}

}
\keyword{internal}
