% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPEDSuploadables.R
\docType{package}
\name{IPEDSuploadables}
\alias{IPEDSuploadables}
\alias{IPEDSuploadables-package}
\title{\code{IPEDSuploadables} package}
\description{
Tools to assist data formatting for upload to IPEDS surveys
}
\details{
See the README on
\href{https://github.com/AlisonLanski/IPEDSuploadables#readme}{GitHub}
}
