% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_adm_part_B.R
\name{make_adm_part_B}
\alias{make_adm_part_B}
\title{Make Admissions Part B (First-time student demographics)}
\usage{
make_adm_part_B(df)
}
\arguments{
\item{This}{function relies on the student data frame that conforms to the documented specifications
for the Admissions module of the IPEDS Uploadables package.}
}
\value{
Part B data prepared for inclusion in the final text file for uploading to the IPEDS portal.
}
\description{
Aggregates First-Time student applicants by sex, race/ethnicity.
}
