% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_adm_part_C.R
\name{make_adm_part_C}
\alias{make_adm_part_C}
\title{Make Admissions Part C (First-time SAT/ACT scores)}
\usage{
make_adm_part_C(df)
}
\arguments{
\item{df}{A dataframe of applicant information}
}
\value{
Admissions Part C data with the required IPEDS structure
}
\description{
If SAT or ACT scores were used for admission decisions it calculates the requisite percentiles for submission.
}
