% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_adm_part_C.R
\name{make_adm_part_C}
\alias{make_adm_part_C}
\title{Make Admissions Part C (First-time SAT/ACT scores)}
\usage{
make_adm_part_C(df)
}
\arguments{
\item{This}{function relies on the student data frame that conforms to the documented specifications
for the Admissions module of the IPEDS Uploadables package.}
}
\value{
Part C data prepared for inclusion in the final text file for uploading to the IPEDS portal.
}
\description{
If SAT or ACT scores were used for admission decisions it calculates the requisite percentiles for submission.
}
