% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_adm_part_D.R
\name{make_adm_part_D}
\alias{make_adm_part_D}
\title{Make Admissions Part D (First-time sex unknown)}
\usage{
make_adm_part_D(df)
}
\arguments{
\item{This}{function relies on the student data frame that conforms to the documented specifications
for the Admissions module of the IPEDS Uploadables package.}
}
\value{
Part D data prepared for inclusion in the final text file for uploading to the IPEDS portal.
}
\description{
Counts the number of unknown sex for First-time students
}
