% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_adm_part_D.R
\name{make_adm_part_D}
\alias{make_adm_part_D}
\title{Make Admissions Part D (First-time sex unknown)}
\usage{
make_adm_part_D(df)
}
\arguments{
\item{df}{A dataframe of applicant information}
}
\value{
Admissions Part D data with the required IPEDS structure
}
\description{
Counts the number of unknown sex for First-time students
}
