% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_adm_part_F.R
\name{make_adm_part_F}
\alias{make_adm_part_F}
\title{Make Admissions Part F (Transfer student demographics)}
\usage{
make_adm_part_F(df)
}
\arguments{
\item{This}{function relies on the student data frame that conforms to the documented specifications
for the Admissions module of the IPEDS Uploadables package.}
}
\value{
Part F data prepared for inclusion in the final text file for uploading to the IPEDS portal.
}
\description{
Aggregates Transfer student applicants by sex, race/ethnicity.
}
