% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_adm_part_F.R
\name{make_adm_part_F}
\alias{make_adm_part_F}
\title{Make Admissions Part F (Transfer student demographics)}
\usage{
make_adm_part_F(df)
}
\arguments{
\item{df}{A dataframe of applicant information}
}
\value{
Admissions Part F data with the required IPEDS structure
}
\description{
Aggregates Transfer student applicants by sex, race/ethnicity.
}
