% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_adm_part_G.R
\name{make_adm_part_G}
\alias{make_adm_part_G}
\title{Make Admissions Part G (Transfer sex unknown)}
\usage{
make_adm_part_G(df)
}
\arguments{
\item{This}{function relies on the student data frame that conforms to the documented specifications
for the Admissions module of the IPEDS Uploadables package.}
}
\value{
Part G data prepared for inclusion in the final text file for uploading to the IPEDS portal.
}
\description{
Counts the number of unknown sex for Transfer students
}
