% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_adm_part_H.R
\name{make_adm_part_H}
\alias{make_adm_part_H}
\title{Make Admissions Part H (Transfer SAT/ACT scores)}
\usage{
make_adm_part_H(df)
}
\arguments{
\item{df}{A dataframe of applicant information}
}
\value{
Admissions Part H data with the required IPEDS structure
}
\description{
If SAT or ACT scores were used for admission decisions it calculates the requisite percentiles for submission.
}
