% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_com_part_A.R
\name{make_com_part_A}
\alias{make_com_part_A}
\title{Make Completions Part A}
\usage{
make_com_part_A(df, extracips = NULL, output = "part")
}
\arguments{
\item{df}{A dataframe of student/degree information}

\item{extracips}{A dataframe of cips offered by the institution but not in \code{'df'}}

\item{output}{A string (\code{"part"}, \code{"full"}, or \code{"both"})}
}
\value{
A text file
}
\description{
Make Completions Part A
}
