% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_com_part_A.R
\name{make_com_part_A}
\alias{make_com_part_A}
\title{Make Completions Part A}
\usage{
make_com_part_A(df, extracips = NULL)
}
\arguments{
\item{df}{A dataframe}

\item{extracips}{A dataframe}
}
\value{
A text file
}
\description{
Make Completions Part A
}
