% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_com_part_B.R
\name{make_com_part_B}
\alias{make_com_part_B}
\title{Make Completions Part B}
\usage{
make_com_part_B(df, extracips = NULL, output = "part")
}
\arguments{
\item{df}{A dataframe of student/degree information}

\item{extracips}{A dataframe of cips offered by the institution but not in \code{'df'}}

\item{output}{A string (\code{"part"}, \code{"full"}, or \code{"both"})}
}
\value{
A text file
}
\description{
Make Completions Part B
}
