% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_com_part_C.R
\name{make_com_part_C}
\alias{make_com_part_C}
\title{Make Completions Part C}
\usage{
make_com_part_C(df, output = "part")
}
\arguments{
\item{df}{A dataframe of student/degree information}

\item{output}{A string (\code{"part"}, \code{"full"}, or \code{"both"})#'}
}
\value{
A text file
}
\description{
Make Completions Part C
}
