% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_com_part_D.R
\name{make_com_part_D}
\alias{make_com_part_D}
\title{Make Completions Part D}
\usage{
make_com_part_D(df, extracips = NULL)
}
\arguments{
\item{df}{A dataframe}

\item{extracips}{A dataframe}
}
\value{
A text file
}
\description{
Make Completions Part D
}
