% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_com_part_D.R
\name{make_com_part_D}
\alias{make_com_part_D}
\title{Make Completions Part D}
\usage{
make_com_part_D(df, extracips = NULL, output = "part", format = "both")
}
\arguments{
\item{df}{A dataframe of student/degree information}

\item{extracips}{A dataframe of cips offered by the institution but not in \code{'df'}}

\item{output}{A string (\code{"part"}, \code{"full"}, or \code{"both"})}

\item{format}{A string (\code{"uploadable"}, \code{"readable"}, or \code{"both"})}
}
\value{
A text file
}
\description{
Make Completions Part D
}
