% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_com_part_E.R
\name{make_com_part_E}
\alias{make_com_part_E}
\title{Make Completions Part E (gender details)}
\usage{
make_com_part_E(df, ugender, ggender)
}
\arguments{
\item{df}{A dataframe of student/degree information}

\item{ugender}{A boolean: TRUE means you are collecting and able to report
"another gender" for undergraduate completers. Set as FALSE if necessary}

\item{ggender}{A boolean: TRUE means you are collecting and able to report
"another gender" for graduate completers. Set as FALSE if necessary}
}
\value{
A text file
}
\description{
Make Completions Part E (gender details)
}
