% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_com_part_E.R
\name{make_com_part_E}
\alias{make_com_part_E}
\title{Make Completions Part E (gender details)}
\usage{
make_com_part_E(df)
}
\arguments{
\item{df}{A dataframe of student/degree information}
}
\value{
A text file
}
\description{
Make Completions Part E (gender details)
}
