% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_e1d_part_A.R
\name{make_e1d_part_A}
\alias{make_e1d_part_A}
\title{Make 12 Month Enrollment Part A}
\usage{
make_e1d_part_A(df)
}
\arguments{
\item{df}{A dataframe of student/degree information}
}
\value{
A text file
}
\description{
Make 12 Month Enrollment Part A
}
