% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_e1d_part_B.R
\name{make_e1d_part_B}
\alias{make_e1d_part_B}
\title{Make 12 Month Enrollment Part B}
\usage{
make_e1d_part_B(df, output = "part", format = "both")
}
\arguments{
\item{df}{A dataframe with summarized credit hours and student information}

\item{output}{A string (\code{"part"}, \code{"full"}, or \code{"both"})}

\item{format}{A string (\code{"uploadable"}, \code{"readable"}, or \code{"both"})}
}
\value{
A text file
}
\description{
Make 12 Month Enrollment Part B
}
