% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_e1d_part_C.R
\name{make_e1d_part_C}
\alias{make_e1d_part_C}
\title{Make 12 Month Enrollment Part C}
\usage{
make_e1d_part_C(df)
}
\arguments{
\item{df}{A dataframe of student/degree information}
}
\value{
A dataframe with the required IPEDS structure for this survey part
}
\description{
Make 12 Month Enrollment Part C
}
