% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_e1d_part_D.R
\name{make_e1d_part_D}
\alias{make_e1d_part_D}
\title{Make 12 Month Enrollment Part D (gender details)}
\usage{
make_e1d_part_D(df, ugender, ggender)
}
\arguments{
\item{df}{A dataframe of student/degree information}

\item{ugender}{A boolean: TRUE means you are collecting and able to report
"another gender" for undergraduate students, even if you have no (or few)
such students. Set as FALSE if necessary}

\item{ggender}{A boolean: TRUE means you are collecting and able to report
"another gender" for graduate students, even if you have no (or few) such
students. Set as FALSE if necessary}
}
\value{
A dataframe with the required IPEDS structure for this survey part
}
\description{
Make 12 Month Enrollment Part D (gender details)
}
