% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ef1_part_A.R
\name{make_ef1_part_A}
\alias{make_ef1_part_A}
\title{Make Fall Enrollment Part A}
\usage{
make_ef1_part_A(df, cips = TRUE, output = "part", format = "both")
}
\arguments{
\item{df}{A dataframe of student information}

\item{cips}{A logical indicating if part A  needs to provide breakdowns by particular CIPs}

\item{output}{A string (\code{"part"}, \code{"full"}, or \code{"both"})}

\item{format}{A string (\code{"uploadable"}, \code{"readable"}, or \code{"both"})}
}
\value{
A text file
}
\description{
Breakdown of students level and demographics; also by designated CIPs in required years
}
