% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ef1_part_A.R
\name{make_ef1_part_A}
\alias{make_ef1_part_A}
\title{Make Fall Enrollment Part A}
\usage{
make_ef1_part_A(df, extracips = NULL, output = "part", format = "both")
}
\arguments{
\item{df}{A dataframe of student/degree information}

\item{extracips}{A dataframe of cips offered by the institution but not in \code{'df'}}

\item{output}{A string (\code{"part"}, \code{"full"}, or \code{"both"})}

\item{format}{A string (\code{"uploadable"}, \code{"readable"}, or \code{"both"})}
}
\value{
A text file
}
\description{
Make Fall Enrollment Part A
}
