% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ef1_part_B.R
\name{make_ef1_part_B}
\alias{make_ef1_part_B}
\title{Make Fall Enrollment Part B}
\usage{
make_ef1_part_B(df)
}
\arguments{
\item{df}{A dataframe of student information}
}
\value{
A dataframe with the required IPEDS structure for this survey part
}
\description{
Student Counts by Age/gender
}
