% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ef1_part_F.R
\name{make_ef1_part_F}
\alias{make_ef1_part_F}
\title{Make Fall Enrollment Part F}
\usage{
make_ef1_part_F(df, output = "part", format = "both")
}
\arguments{
\item{df}{A dataframe of student/degree information}

\item{output}{A string (\code{"part"}, \code{"full"}, or \code{"both"})}

\item{format}{A string (\code{"uploadable"}, \code{"readable"}, or \code{"both"})}
}
\value{
A text file
}
\description{
Make Fall Enrollment Part F
}
