% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gr_part_B.R
\name{make_gr_part_B}
\alias{make_gr_part_B}
\title{Make Graduation Rates Part B}
\usage{
make_gr_part_B(df)
}
\arguments{
\item{df}{A dataframe of student/degree information}
}
\value{
A dataframe with the required IPEDS structure for this survey part
}
\description{
Make Graduation Rates Part B
}
