% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gr_part_C.R
\name{make_gr_part_C}
\alias{make_gr_part_C}
\title{Make Graduation Rates Part C}
\usage{
make_gr_part_C(df)
}
\arguments{
\item{df}{A dataframe of student/degree information}
}
\value{
A dataframe with the required IPEDS structure for this survey part
}
\description{
Make Graduation Rates Part C
}
