% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gr_part_E.R
\name{make_gr_part_E}
\alias{make_gr_part_E}
\title{Make Graduation Rates Part E (gender details)}
\usage{
make_gr_part_E(df, ugender)
}
\arguments{
\item{df}{A dataframe of student/degree information for unduplicated
undergraduates}

\item{ugender}{A boolean: TRUE means you are collecting and able to report
"another gender" for undergraduate students, even if you have no (or few)
such students. Set as FALSE if necessary. Argument deprecated from the produce function.}
}
\value{
A dataframe with the required IPEDS structure for this survey part
}
\description{
`r lifecycle::badge("deprecated")`

 This section was removed by IPEDS in the 2024-2025 reporting cycle, after
 being required for the 2023-2024 reporting cycle. It remains here in case it
 is re-added in a future year. I have updated testing/produce code so nothing
 else calls it. By making it internal, it won't be listed but remains
 available.
}
\keyword{internal}
