% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hr_part_A1.R
\name{make_hr_part_A1}
\alias{make_hr_part_A1}
\title{Produce IPEDS HR Part A1}
\usage{
make_hr_part_A1(df, output = "part")
}
\arguments{
\item{df}{a dataframe}

\item{output}{a string (\code{"part"}, \code{"full"}, or \code{"both"})}
}
\value{
a txt file
}
\description{
Part A1 --- Count of FT Instructional staff by tenure status, academic rank, and race/ethnicity/gender
}
