% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hr_part_A1.R
\name{make_hr_part_A1}
\alias{make_hr_part_A1}
\title{Make Human Resources Part A1}
\usage{
make_hr_part_A1(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
a txt file
}
\description{
Part A1 --- COUNT of FT INSTRUCTIONAL staff by tenure status, academic rank, and race/ethnicity/gender
}
