% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hr_part_B1.R
\name{make_hr_part_B1}
\alias{make_hr_part_B1}
\title{Produce IPEDS HR Part B1}
\usage{
make_hr_part_B1(df, output = "part")
}
\arguments{
\item{df}{a dataframe}

\item{output}{a string (\code{"part"}, \code{"full"}, or \code{"both"})}
}
\value{
a txt file
}
\description{
HR Part B1 --- Count of FT Non-instructional staff by occupational category
}
