% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hr_part_B1.R
\name{make_hr_part_B1}
\alias{make_hr_part_B1}
\title{Make Human Resources Part B1}
\usage{
make_hr_part_B1(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
a txt file
}
\description{
HR Part B1 --- COUNT of FT Non-instructional staff by occupational category
}
