% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hr_part_B2.R
\name{make_hr_part_B2}
\alias{make_hr_part_B2}
\title{Produce IPEDS HR Part B2}
\usage{
make_hr_part_B2(df, output = "part", format = "both")
}
\arguments{
\item{df}{a dataframe}

\item{output}{a string (\code{"part"}, \code{"full"}, or \code{"both"})}

\item{format}{A string (\code{"uploadable"}, \code{"readable"}, or \code{"both"})}
}
\value{
a txt file
}
\description{
Part B2 --- Full-time non-instructional staff by tenure, medical school, and occupational category
}
