% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hr_part_D1.R
\name{make_hr_part_D1}
\alias{make_hr_part_D1}
\title{Make Human Resources Part D1}
\usage{
make_hr_part_D1(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
a txt file
}
\description{
Part D1 --- Part-time staff by occupational category
}
