% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hr_part_D4.R
\name{make_hr_part_D4}
\alias{make_hr_part_D4}
\title{Produce IPEDS HR Part D4}
\usage{
make_hr_part_D4(df, output = "part", format = "both")
}
\arguments{
\item{df}{a dataframe}

\item{output}{a string (\code{"part"}, \code{"full"}, or \code{"both"})}

\item{format}{A string (\code{"uploadable"}, \code{"readable"}, or \code{"both"})}
}
\value{
a txt file
}
\description{
Part D4 --- Part-time Non-instructional staff by medical school, and occupational category
}
