% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hr_part_H2.R
\name{make_hr_part_H2}
\alias{make_hr_part_H2}
\title{Make Human Resources Part H2}
\usage{
make_hr_part_H2(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
A dataframe with the required IPEDS structure for this survey part
}
\description{
Part H2 --- New hires by occupational category, Race/Ethnicity/Gender
}
