% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_om_part_A.R
\name{make_om_part_A}
\alias{make_om_part_A}
\title{Make Outcome Measures Part A}
\usage{
make_om_part_A(df)
}
\arguments{
\item{df}{A dataframe of student statuses}
}
\value{
A text file ready for IPEDS upload
}
\description{
Establishing the Outcome Measures cohorts
}
