% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_om_part_B.R
\name{make_om_part_B}
\alias{make_om_part_B}
\title{Make Outcome Measures Part B}
\usage{
make_om_part_B(df, output = "part", format = "both")
}
\arguments{
\item{df}{A dataframe of student statuses}

\item{output}{A string (\code{"part"}, \code{"full"}, or \code{"both"})}

\item{format}{A string (\code{"uploadable"}, \code{"readable"}, or \code{"both"})}
}
\value{
A text file ready for IPEDS upload
}
\description{
Award Status at Four Years after Entry
}
\examples{

}
