% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_om_part_C.R
\name{make_om_part_C}
\alias{make_om_part_C}
\title{Make Outcome Measures Part C}
\usage{
make_om_part_C(df)
}
\arguments{
\item{df}{A dataframe of student statuses}
}
\value{
A text file ready for IPEDS upload
}
\description{
Award Status at Six Years after Entry
}
