% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_om_part_D.R
\name{make_om_part_D}
\alias{make_om_part_D}
\title{Make Outcome Measures Part D}
\usage{
make_om_part_D(df)
}
\arguments{
\item{df}{A dataframe of student statuses}
}
\value{
A text file ready for IPEDS upload
}
\description{
Award Status and Enrollment at Eight Years after Entry
}
