% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_adm_data_frame.R
\name{prep_adm_data_frame}
\alias{prep_adm_data_frame}
\title{Some initial recodes for Admissions parts C and D}
\usage{
prep_adm_data_frame(df)
}
\arguments{
\item{df}{a dataframe of applicant level data}
}
\description{
Some initial recodes for Admissions parts C and D
}
