% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_com_data_frame.R
\name{prep_com_data_frame}
\alias{prep_com_data_frame}
\title{Some initial recodes for Completions}
\usage{
prep_com_data_frame(df)
}
\arguments{
\item{df}{a dataframe of student level data or cip information}
}
\value{
df
}
\description{
Some initial recodes for Completions
}
