% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_hr_data_frame.R
\name{prep_hr_data_frame}
\alias{prep_hr_data_frame}
\title{Some initial recoding for Human Resources}
\usage{
prep_hr_data_frame(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
A dataframe ready for the make_hr scripts
}
\description{
Some initial recoding for Human Resources
}
