% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_com_report.R
\name{produce_com_report}
\alias{produce_com_report}
\title{Shortcut function to do all steps to produce a report}
\usage{
produce_com_report(df, extracips = NULL, part = "all")
}
\arguments{
\item{df}{A dataframe set up according to the readme}

\item{extracips}{A dataframe set up according to the readme (optional)}

\item{part}{A string with what part of the reprt you want to produce: 'all', 'A', etc.}
}
\value{
A txt file at the path of your choice
}
\description{
Shortcut function to do all steps to produce a report
}
