% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_ef1_report.R
\name{produce_ef1_report}
\alias{produce_ef1_report}
\title{Shortcut function with all steps to provide a Fall Enrollment report}
\usage{
produce_ef1_report(
  students,
  retention,
  part = "ALL",
  include_optional = FALSE,
  format = "uploadable"
)
}
\arguments{
\item{students}{A dataframe set up according to the readme with student data}

\item{retention}{A dataframe set up according to the readme with retention data}

\item{part}{A string with what part of the report you want to produce: 'all', 'A', etc.}

\item{include_optional}{A boolean flag for whether optional parts should be included}

\item{format}{A string (\code{"uploadable"} will produce a properly formatted
upload file. \code{"readable"} will produce a csv of the upload file (only
works for one part at a time). \code{"both"} will provide both options, but
only works with one part at a time.}
}
\value{
A txt or csv file at the path of your choice
}
\description{
Shortcut function with all steps to provide a Fall Enrollment report
}
\examples{
\dontshow{
.old_wd <- setwd(tempdir())
}
#entire report
produce_ef1_report(ef1_students, ef1_retention)
#entire report with optional sections
produce_ef1_report(ef1_students, ef1_retention, include_optional = TRUE)
#one part as csv instead of key-value
produce_ef1_report(ef1_students, part = 'D', format = 'readable')
\dontshow{
setwd(.old_wd)
}
}
