% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_gr200_report.R
\name{produce_gr200_report}
\alias{produce_gr200_report}
\title{Grad Rates 200 wrapper function to produce report}
\usage{
produce_gr200_report(df, format = "uploadable")
}
\arguments{
\item{df}{a dataframe set up according to the readme}

\item{format}{A string (\code{"uploadable"} will produce a properly formatted
upload file. \code{"readable"} will produce a csv of the upload file (only
works for one part at a time). \code{"both"} will provide both options, but
only works with one part at a time.}
}
\value{
A txt or csv file at the path of your choice
}
\description{
Grad Rates 200 wrapper function to produce report
}
