% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_gr_report.R
\name{produce_gr_report}
\alias{produce_gr_report}
\title{Shortcut function with all steps to provide a Graduation Rates report}
\usage{
produce_gr_report(df, part = "ALL", format = "uploadable", ugender = TRUE)
}
\arguments{
\item{df}{a dataframe set up according to the readme}

\item{part}{a string with what part of the report you want to produce "all",
"A1", etc.}

\item{format}{A string (\code{"uploadable"} will produce a properly formatted
upload file. \code{"readable"} will produce a csv of the upload file (only
works for one part at a time). \code{"both"} will provide both options, but
only works with one part at a time.}

\item{ugender}{A boolean: TRUE means you are collecting and able to report
"another gender" for undergraduate students, even if you have no (or few)
such students. Set as FALSE if necessary}
}
\value{
A txt or csv file at the path of your choice
}
\description{
Shortcut function with all steps to provide a Graduation Rates report
}
\examples{
\dontshow{
#set temp directory for this example (not necessary for users)
.old_wd <- setwd(tempdir())
}
\donttest{
#entire report
produce_gr_report(gr_students)

#one part in csv format instead of key-value
produce_gr_report(gr_students, part = "B", format = "readable")
}
\dontshow{
#reset directory for this example (not necessary for users)
setwd(.old_wd)
}
}
