% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_gr_report.R
\name{produce_gr_report}
\alias{produce_gr_report}
\title{Shortcut function to do all steps to produce a report}
\usage{
produce_gr_report(df, part = "ALL", format = "uploadable")
}
\arguments{
\item{df}{a dataframe set up according to the readme}

\item{part}{a string with what part of the report you want to produce "all", "A1", etc.}

\item{format}{A string (\code{"uploadable"} will produce a properly formatted
upload file. \code{"readable"} will produce a csv of the upload file (only
works for one part at a time). \code{"both"} will provide both options, but
only works with one part at a time.}
}
\value{
A txt or csv file at the path of your choice
}
\description{
Shortcut function to do all steps to produce a report
}
