% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_om_report.R
\name{produce_om_report}
\alias{produce_om_report}
\title{Shortcut function to do all steps to produce a report}
\usage{
produce_om_report(df, part = "ALL")
}
\arguments{
\item{df}{A dataframe set up according to the readme}

\item{part}{A string with what part of the report you want to produce: 'all', 'A', etc.}
}
\value{
A txt file at the path of your choice
}
\description{
Shortcut function to do all steps to produce a report
}
