% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_report.R
\name{write_report}
\alias{write_report}
\title{Write the prepared data to a txt file in IPEDS format}
\usage{
write_report(
  df,
  component,
  part,
  output,
  append = FALSE,
  format = "uploadable"
)
}
\arguments{
\item{df}{a dataframe (prepared via the 'make' scripts)}

\item{component}{a string (which IPEDS survey)}

\item{part}{a string (which upload part of the survey)}

\item{output}{a string (wanting to write just this part, the entire report, or both)}

\item{append}{a logical (if the txt file should append this part (T) or overwrite the entire file(F))}

\item{format}{A string (\code{"uploadable"}, \code{"readable"}, or \code{"both"})}
}
\value{
a txt file (at the path location)
}
\description{
Write the prepared data to a txt file in IPEDS format
}
